requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cycler==0.10.0
decorator==4.4.2
einops==0.2.0
filelock==3.0.12
Flask==2.0.1
gdown==3.11.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lmdb==1.0.0
MarkupSafe==2.0.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
opencv-python==4.5.3.56
Pillow>=8.3.2
prefetch-generator==1.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.7.1
six==1.16.0
tifffile==2021.6.6
torch==1.8.2
torchaudio==0.8.2
torchvision==0.9.2
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
vietocr==0.3.2
Werkzeug==2.0.1
